pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "yajurvedj", email = "yjayavar@uab.edu" }
]
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.3.5",
    "html2text>=2025.4.15",
    "langchain>=1.1.0",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-openai>=1.1.0",
    "langsmith>=0.4.53",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.52.1",
]

[project.scripts]
rag = "rag:main"

[build-system]
requires = ["uv_build>=0.9.8,<0.10.0"]
build-backend = "uv_build"
